+++ .vscode/settings.json
@@ -49,6 +49,5 @@
   },
   "files.insertFinalNewline": true,
   "nixEnvSelector.nixFile": "${workspaceRoot}/flake.nix",
-  "testing.automaticallyOpenPeekView": "never",
   "files.eol": "\n",
 }
