+++ packages/cluster-workflow/test/Workflow.test.ts
@@ -1,9 +1,11 @@
 import * as Activity from "@effect/cluster-workflow/Activity"
+import * as ActivityContext from "@effect/cluster-workflow/ActivityContext"
 import * as CrashableRuntime from "@effect/cluster-workflow/CrashableRuntime"
 import * as DurableExecutionJournal from "@effect/cluster-workflow/DurableExecutionJournal"
 import * as DurableExecutionJournalInMemory from "@effect/cluster-workflow/DurableExecutionJournalInMemory"
 import * as utils from "@effect/cluster-workflow/test/utils"
 import * as Workflow from "@effect/cluster-workflow/Workflow"
+import * as WorkflowRunner from "@effect/cluster-workflow/WorkflowRunner"
 import * as Schema from "@effect/schema/Schema"
 import * as Deferred from "effect/Deferred"
 import * as Effect from "effect/Effect"
@@ -24,6 +26,12 @@ describe.concurrent("Workflow", () => {
       Logger.withMinimumLogLevel(LogLevel.Debug)
     )
 
+  class StartWorkflowRequest
+    extends Schema.TaggedRequest<StartWorkflowRequest>()("StartWorkflow", Schema.string, Schema.number, {
+      executionId: Schema.string
+    })
+  {}
+
   it("Should run as expected if not crashed", () => {
     return Effect.gen(function*(_) {
       const mocked = utils.mockEffect(() => Exit.succeed(42))
@@ -33,8 +41,10 @@ describe.concurrent("Workflow", () => {
         Activity.make("activity", Schema.never, Schema.number)
       )
 
+      const workflow = Workflow.make(StartWorkflowRequest, (_) => _.executionId, () => activity)
+
       const exit = yield* _(
-        Workflow.unsafeAttempt("wf", Schema.never, Schema.number)(activity),
+        WorkflowRunner.resume(workflow)(new StartWorkflowRequest({ executionId: "wf" })),
         Effect.exit
       )
 
@@ -52,13 +62,48 @@ describe.concurrent("Workflow", () => {
         Activity.make("activity", Schema.string, Schema.number)
       )
 
+      const workflow = Workflow.make(StartWorkflowRequest, (_) => _.executionId, () => activity)
+
       const exit = yield* _(
-        Workflow.unsafeAttempt("wf", Schema.string, Schema.number)(activity),
+        WorkflowRunner.resume(workflow)(new StartWorkflowRequest({ executionId: "wf" })),
         Effect.exit
       )
 
       expect(mocked.spy).toHaveBeenCalledOnce()
       expect(exit).toEqual(Exit.fail("error"))
+    }).pipe(withTestEnv, Effect.runPromise)
+  })
+
+  it("Should be able to compose different workflows", () => {
+    const firstActivity = utils.mockActivity("activity1", Schema.never, Schema.number, () => Exit.succeed(1))
+    const secondActivity = utils.mockActivity("activity2", Schema.never, Schema.number, () => Exit.succeed(2))
+
+    class FirstWorkflow extends Schema.TaggedRequest<FirstWorkflow>()("FirstWorkflow", Schema.never, Schema.number, {
+      id: Schema.string
+    }) {}
+
+    const firstWorkflow = Workflow.make(FirstWorkflow, (_) => _.id, () => firstActivity.activity)
+
+    class SecondWorkflow extends Schema.TaggedRequest<SecondWorkflow>()("SecondWorkflow", Schema.never, Schema.number, {
+      id: Schema.string
+    }) {}
+
+    const secondWorkflow = Workflow.make(SecondWorkflow, (_) => _.id, () => secondActivity.activity)
+
+    const mergedWorkflow = Workflow.union(firstWorkflow, secondWorkflow)
+
+    return Effect.gen(function*(_) {
+      const exit = yield* _(WorkflowRunner.resume(mergedWorkflow)(new FirstWorkflow({ id: "wf1" })), Effect.exit)
+
+      expect(exit).toEqual(Exit.succeed(1))
+      expect(firstActivity.spy).toHaveBeenCalledOnce()
+      expect(secondActivity.spy).not.toHaveBeenCalled()
+
+      const exit2 = yield* _(WorkflowRunner.resume(mergedWorkflow)(new SecondWorkflow({ id: "wf2" })), Effect.exit)
+
+      expect(exit2).toEqual(Exit.succeed(2))
+      expect(firstActivity.spy).toHaveBeenCalledOnce()
+      expect(secondActivity.spy).toHaveBeenCalledOnce()
     }).pipe(withTestEnv, Effect.runPromise)
   })
 
@@ -68,14 +113,14 @@ describe.concurrent("Workflow", () => {
 
       const activity = pipe(mocked.effect, Activity.make("activity", Schema.never, Schema.number))
 
-      const workflow = Workflow.unsafeAttempt("wf", Schema.never, Schema.number)(activity)
+      const workflow = Workflow.make(StartWorkflowRequest, (_) => _.executionId, () => activity)
 
       const exit1 = yield* _(
-        workflow,
+        WorkflowRunner.resume(workflow)(new StartWorkflowRequest({ executionId: "wf" })),
         Effect.exit
       )
       const exit2 = yield* _(
-        workflow,
+        WorkflowRunner.resume(workflow)(new StartWorkflowRequest({ executionId: "wf" })),
         Effect.exit
       )
 
@@ -90,21 +135,21 @@ describe.concurrent("Workflow", () => {
       const mockedUse = utils.mockActivity("use", Schema.never, Schema.number, () => Exit.succeed(2))
       const mockedRelease = utils.mockActivity("release", Schema.never, Schema.number, () => Exit.succeed(3))
 
-      const workflow = (shouldCrash: boolean) =>
-        CrashableRuntime.runWithCrash((crash) =>
-          pipe(
+      const executeAttempt = (shouldCrash: boolean) =>
+        CrashableRuntime.runWithCrash((crash) => {
+          const workflow = Workflow.make(StartWorkflowRequest, (_) => _.executionId, () =>
             Effect.acquireUseRelease(
               mockedAcquire.activity,
               () => mockedUse.activityWithBody(shouldCrash ? crash : Effect.succeed(2)),
               () => mockedRelease.activity
-            ),
-            Workflow.unsafeAttempt("wf", Schema.never, Schema.number)
-          )
-        )
+            ))
+
+          return WorkflowRunner.resume(workflow)(new StartWorkflowRequest({ executionId: "wf" }))
+        })
 
       // first execution should crash
       yield* _(
-        workflow(true),
+        executeAttempt(true),
         Effect.exit
       )
 
@@ -114,7 +159,7 @@ describe.concurrent("Workflow", () => {
 
       // now the workflow gets executed again without crashing
       yield* _(
-        workflow(false),
+        executeAttempt(false),
         Effect.exit
       )
 
@@ -130,21 +175,21 @@ describe.concurrent("Workflow", () => {
       const mockedUse = utils.mockActivity("use", Schema.never, Schema.number, () => Exit.succeed(2))
       const mockedRelease = utils.mockActivity("release", Schema.never, Schema.number, () => Exit.succeed(3))
 
-      const workflow = (shouldCrash: boolean) =>
-        CrashableRuntime.runWithCrash((crash) =>
-          pipe(
+      const executeAttempt = (shouldCrash: boolean) =>
+        CrashableRuntime.runWithCrash((crash) => {
+          const workflow = Workflow.make(StartWorkflowRequest, (_) => _.executionId, () =>
             Effect.acquireUseRelease(
               mockedAcquire.activity,
               () => mockedUse.activity,
-              () => mockedRelease.activityWithBody(shouldCrash ? crash : Effect.succeed(2))
-            ),
-            Workflow.unsafeAttempt("wf", Schema.never, Schema.number)
-          )
-        )
+              () => mockedRelease.activityWithBody(shouldCrash ? crash : Effect.succeed(3))
+            ))
+
+          return WorkflowRunner.resume(workflow)(new StartWorkflowRequest({ executionId: "wf" }))
+        })
 
       // first execution should crash
       yield* _(
-        workflow(true),
+        executeAttempt(true),
         Effect.exit
       )
 
@@ -154,7 +199,7 @@ describe.concurrent("Workflow", () => {
 
       // now the workflow gets executed again without crashing
       yield* _(
-        workflow(false),
+        executeAttempt(false),
         Effect.exit
       )
 
@@ -170,8 +215,8 @@ describe.concurrent("Workflow", () => {
 
       yield* _(
         CrashableRuntime.retryWhileCrashes((runtime) =>
-          runtime.run(() =>
-            pipe(
+          runtime.run(() => {
+            const activity = pipe(
               Activity.currentAttempt,
               Effect.flatMap((attempt) =>
                 pipe(
@@ -188,10 +233,13 @@ describe.concurrent("Workflow", () => {
                   )
                 )
               ),
-              Activity.make("activity", Schema.never, Schema.number),
-              Workflow.unsafeAttempt("wf", Schema.never, Schema.number)
+              Activity.make("activity", Schema.never, Schema.number)
             )
-          )
+
+            const workflow = Workflow.make(StartWorkflowRequest, (_) => _.executionId, () => activity)
+
+            return WorkflowRunner.resume(workflow)(new StartWorkflowRequest({ executionId: "wf" }))
+          })
         )
       )
 
@@ -207,16 +255,21 @@ describe.concurrent("Workflow", () => {
       const mockedRelease = utils.mockEffect(() => Exit.succeed(1))
       const mockedUse = utils.mockEffect(() => Exit.succeed(1))
       const latch = yield* _(Deferred.make<never, void>())
+      const persistenceIdRef = yield* _(Ref.make(""))
 
       const exit = yield* _(
         Effect.acquireUseRelease(Effect.succeed(1), () =>
           pipe(
             mockedUse.effect,
+            Effect.zipRight(
+              Effect.flatMap(ActivityContext.ActivityContext, (ctx) => Ref.set(persistenceIdRef, ctx.persistenceId))
+            ),
             Effect.zipRight(Effect.never),
             Effect.zipLeft(Deferred.succeed(latch, undefined), { concurrent: true })
           ), () => mockedRelease.effect),
         mockedActivity.activityWithBody,
-        Workflow.unsafeAttempt("wf", Schema.never, Schema.number),
+        (activity) => Workflow.make(StartWorkflowRequest, (_) => _.executionId, () => activity),
+        (workflow) => WorkflowRunner.resume(workflow)(new StartWorkflowRequest({ executionId: "wf" })),
         Effect.forkDaemon,
         Effect.tap(Deferred.await(latch)),
         Effect.tap((fiber) => Fiber.interrupt(fiber)),
@@ -228,8 +281,9 @@ describe.concurrent("Workflow", () => {
         Stream.runCount
       )
 
+      const persistenceId = yield* _(Ref.get(persistenceIdRef))
       const activityJournalEntryCount = yield* _(
-        DurableExecutionJournal.read("wf__", Schema.never, Schema.number),
+        DurableExecutionJournal.read(persistenceId, Schema.never, Schema.number),
         Stream.runCount
       )
